Refactor: move Campaign.kt to data.models
