+++ app/src/main/java/fr/free/nrw/commons/campaigns/CampaignView.java
@@ -10,6 +10,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import fr.free.nrw.commons.data.models.Campaign;
 import fr.free.nrw.commons.theme.BaseActivity;
 import org.wikipedia.util.DateUtil;
