+++ app/src/main/java/fr/free/nrw/commons/campaigns/CampaignsPresenter.java
@@ -2,6 +2,7 @@ package fr.free.nrw.commons.campaigns;
 
 import android.annotation.SuppressLint;
 
+import fr.free.nrw.commons.data.models.Campaign;
 import java.text.ParseException;
 import java.util.Collections;
 import java.util.Date;
