+++ app/src/main/java/fr/free/nrw/commons/campaigns/ICampaignsView.java
@@ -1,6 +1,7 @@
 package fr.free.nrw.commons.campaigns;
 
 import fr.free.nrw.commons.MvpView;
+import fr.free.nrw.commons.data.models.Campaign;
 
 /**
  * Interface which defines the view contracts of the campaign view
