+++ app/src/test/kotlin/fr/free/nrw/commons/campaigns/CampaignsPresenterTest.kt
@@ -1,6 +1,7 @@
 package fr.free.nrw.commons.campaigns
 
 import com.nhaarman.mockitokotlin2.verify
+import fr.free.nrw.commons.data.models.Campaign
 import fr.free.nrw.commons.mwapi.OkHttpJsonApiClient
 import io.reactivex.Single
 import io.reactivex.disposables.Disposable
